+++ code/nel/src/3d/driver/direct3d/driver_direct3d.h
@@ -839,7 +839,7 @@ public:
 	// ***************************************************************************
 
 	// Mode initialisation, requests
-	virtual bool			init (uint windowIcon = 0, emptyProc exitFunc = 0);
+	virtual bool			init (uintptr_t windowIcon = 0, emptyProc exitFunc = 0);
 	virtual bool			setDisplay(nlWindow wnd, const GfxMode& mode, bool show, bool resizeable) throw(EBadDisplay);
 	virtual bool			release();
 	virtual bool			setMode(const GfxMode& mode);
@@ -953,7 +953,7 @@ public:
 	virtual void			setSwapVBLInterval(uint interval);
 	virtual uint			getSwapVBLInterval();
 	virtual void			swapTextureHandle(ITexture &tex0, ITexture &tex1);
-	virtual	uint			getTextureHandle(const ITexture&tex);
+	virtual	uintptr_t		getTextureHandle(const ITexture&tex);
 
 	// Matrix, viewport and frustum
 	virtual void			setFrustum(float left, float right, float bottom, float top, float znear, float zfar, bool perspective = true);
@@ -1893,7 +1893,7 @@ public:
 		H_AUTO_D3D(CDriverD3D_setSamplerState);
 		nlassert (_DeviceInterface);
 		nlassert (sampler<MaxSampler);
-		nlassert (samplerState<MaxSamplerState);
+		nlassert ((int)samplerState<(int)MaxSamplerState);
 
 		// Ref on the state
 		CSamplerState &_samplerState = _SamplerStateCache[sampler][samplerState];
@@ -2025,7 +2025,7 @@ public:
 
 		// Remap high matrices indexes
 		type = (D3DTRANSFORMSTATETYPE)remapMatrixIndex (type);
-		nlassert (type<MaxMatrixState);
+		nlassert ((int)type<(int)MaxMatrixState);
 
 		CMatrixState &theMatrix = _MatrixCache[type];
 #ifdef NL_D3D_USE_RENDER_STATE_CACHE
